refactor: move app logic to server sub-package, embed migrations, split tests
- Extract all container logic into container/server package to allow
  integration tests to import server.NewApp / server.MigrateDatabase
- Embed SQL migrations with //go:embed so binary is self-contained
- Split monolithic integration_test.go into topic files under container/tests/
- Move subnet unit tests to container/server/subnet_test.go
- Update Dockerfile, Makefile, and CI workflow accordingly
