+++ .github/workflows/test.yml
@@ -55,5 +55,5 @@ jobs:
         with:
           go-version: "1.26"
       - name: Integration tests (testcontainers)
-        run: go test -tags=integration -timeout=10m ./...
+        run: go test -tags=integration -timeout=10m ./tests/...
         working-directory: container
