+++ Makefile
@@ -33,7 +33,7 @@ test: ## Run unit tests (container + provider)
 	cd $(REPO_ROOT)/provider  && go test ./...
 
 test-integration: ## Run integration tests (requires Docker)
-	cd $(REPO_ROOT)/container && go test -tags=integration -timeout=10m ./...
+	cd $(REPO_ROOT)/container && go test -tags=integration -timeout=10m ./tests/...
 
 ## ── Local development ───────────────────────────────────────────────────────
