+++ container/main.go
@@ -17,73 +17,23 @@ package main
 
 import (
 	"context"
-	"database/sql"
 	"fmt"
 	"log/slog"
 	"os"
 	"strconv"
 
+	"github.com/boozt-platform/ipam-autopilot/container/server"
 	"github.com/go-sql-driver/mysql"
-	"github.com/gofiber/fiber/v2"
+	"database/sql"
 )
-
-var db *sql.DB
-
-func newApp(database *sql.DB) *fiber.App {
-	db = database
-
-	app := fiber.New()
-	app.Use(requestIDMiddleware())
-	app.Use(tracingMiddleware())
-	app.Use(accessLogMiddleware())
-
-	app.Get("/", func(c *fiber.Ctx) error {
-		return c.SendString("IPAM Autopilot up and running!")
-	})
-
-	// Terraform provider registry — versioned by terraform protocol, kept as-is
-	app.Get("/.well-known/terraform.json", GetTerraformDiscovery)
-	app.Get("/terraform/providers/v1/ipam-autopilot/ipam/versions", GetTerraformVersions)
-	app.Get("/terraform/providers/v1/ipam-autopilot/ipam/:version/download/:os/:arch", GetTerraformVersionDownload)
-
-	// API v1
-	v1 := app.Group("/api/v1")
-	v1.Post("/ranges", CreateNewRange)
-	v1.Get("/ranges", GetRanges)
-	v1.Get("/ranges/:id", GetRange)
-	v1.Delete("/ranges/:id", DeleteRange)
-
-	v1.Get("/domains", GetRoutingDomains)
-	v1.Get("/domains/:id", GetRoutingDomain)
-	v1.Put("/domains/:id", UpdateRoutingDomain)
-	v1.Post("/domains", CreateRoutingDomain)
-	v1.Delete("/domains/:id", DeleteRoutingDomain)
-
-	v1.Get("/audit", GetAuditLogs)
-
-	// Legacy routes — kept for Terraform provider backward compatibility
-	// TODO: remove after provider is updated to use /api/v1
-	app.Post("/ranges", CreateNewRange)
-	app.Get("/ranges", GetRanges)
-	app.Get("/ranges/:id", GetRange)
-	app.Delete("/ranges/:id", DeleteRange)
-
-	app.Get("/domains", GetRoutingDomains)
-	app.Get("/domains/:id", GetRoutingDomain)
-	app.Put("/domains/:id", UpdateRoutingDomain)
-	app.Post("/domains", CreateRoutingDomain)
-	app.Delete("/domains/:id", DeleteRoutingDomain)
-
-	return app
-}
 
 func main() {
 	ctx := context.Background()
 
-	logger := initLogger()
+	logger := server.InitLogger()
 	slog.SetDefault(logger)
 
-	shutdownTracer, err := initTracer(ctx)
+	shutdownTracer, err := server.InitTracer(ctx)
 	if err != nil {
 		slog.Error("failed to initialize tracer", "error", err)
 		os.Exit(1)
@@ -100,7 +50,7 @@ func main() {
 		AllowNativePasswords: true,
 	}
 
-	db, err = sql.Open("mysql", cfg.FormatDSN())
+	db, err := sql.Open("mysql", cfg.FormatDSN())
 	if err != nil {
 		slog.Error("failed to open database", "error", err)
 		os.Exit(1)
@@ -110,13 +60,13 @@ func main() {
 	defer db.Close()
 
 	if os.Getenv("DISABLE_DATABASE_MIGRATION") != "TRUE" {
-		if err = MigrateDatabase(os.Getenv("DATABASE_NAME"), db); err != nil {
+		if err = server.MigrateDatabase(os.Getenv("DATABASE_NAME"), db); err != nil {
 			slog.Error("failed to migrate database", "error", err)
 			os.Exit(1)
 		}
 	}
 
-	app := newApp(db)
+	app := server.NewApp(db)
 
 	port := int64(8080)
 	if os.Getenv("PORT") != "" {
